+++ src/zl-util/ZLMemStream.cpp
@@ -238,7 +238,7 @@ void ZLMemStream::SetChunkSize ( size_t chunkSize ) {
 //----------------------------------------------------------------//
 ZLResultCode ZLMemStream::SetCursor ( long offset ) {
 
-	if (( offset < 0 ) || ( this->mLength < offset )) return ZL_ERROR;
+	if (( offset < 0 ) || ( this->mLength < ( size_t )offset )) return ZL_ERROR;
 
 	this->mCursor = offset;
 	return ZL_OK;
